// src/parser/cst-types/tree-sitter-elm.d.ts
export class Parser {
  parse(input: string | Input, oldTree?: Tree, options?: Options): Tree;
  getIncludedRanges(): Range[];
  getTimeoutMicros(): number;
  setTimeoutMicros(timeout: number): void;
  reset(): void;
  getLanguage(): any;
  setLanguage(language?: any): void;
  getLogger(): Logger;
  setLogger(logFunc?: Logger | false | null): void;
  printDotGraphs(enabled?: boolean, fd?: number): void;
}

export type Options = {
  bufferSize?: number;
  includedRanges?: Range[];
};

export type Point = {
  row: number;
  column: number;
};

export type Range = {
  startIndex: number;
  endIndex: number;
  startPosition: Point;
  endPosition: Point;
};

export type Edit = {
  startIndex: number;
  oldEndIndex: number;
  newEndIndex: number;
  startPosition: Point;
  oldEndPosition: Point;
  newEndPosition: Point;
};

export type Logger = (
  message: string,
  params: { [param: string]: string },
  type: "parse" | "lex"
) => void;

export interface Input {
  (index: number, position?: Point): string | null;
}
// ...
export interface TreeCursor {
  nodeType: string;
  nodeTypeId: number;
  nodeStateId: number;
  nodeText: string;
  nodeIsNamed: boolean;
  nodeIsMissing: boolean;
  startPosition: Point;
  endPosition: Point;
  startIndex: number;
  endIndex: number;
  readonly currentNode: SyntaxNode;
  readonly currentFieldName: string;
  readonly currentFieldId: number;
  readonly currentDepth: number;
  readonly currentDescendantIndex: number;

  reset(node: SyntaxNode): void;
  resetTo(cursor: TreeCursor): void;
  gotoParent(): boolean;
  gotoFirstChild(): boolean;
  gotoLastChild(): boolean;
  gotoFirstChildForIndex(goalIndex: number): boolean;
  gotoFirstChildForPosition(goalPosition: Point): boolean;
  gotoNextSibling(): boolean;
  gotoPreviousSibling(): boolean;
  gotoDescendant(goalDescendantIndex: number): void;
}

export interface Tree {
  readonly rootNode: SyntaxNode;

  rootNodeWithOffset(offsetBytes: number, offsetExtent: Point): SyntaxNode;
  edit(edit: Edit): Tree;
  walk(): TreeCursor;
  getChangedRanges(other: Tree): Range[];
  getIncludedRanges(): Range[];
  getEditedRange(other: Tree): Range;
  printDotGraph(fd?: number): void;
}
// ...
interface NamedNodeBase extends SyntaxNodeBase {
  isNamed: true;
}
// ...
export const enum SyntaxType {
  ERROR = "ERROR",
  AnonymousFunctionExpr = "anonymous_function_expr",
  AnythingPattern = "anything_pattern",
  AsClause = "as_clause",
  BinOpExpr = "bin_op_expr",
  BlockComment = "block_comment",
  CaseOfBranch = "case_of_branch",
  CaseOfExpr = "case_of_expr",
  CharConstantExpr = "char_constant_expr",
  ConsPattern = "cons_pattern",
  Dot = "dot",
  ExposedOperator = "exposed_operator",
  ExposedType = "exposed_type",
  ExposedUnionConstructors = "exposed_union_constructors",
  ExposedValue = "exposed_value",
  ExposingList = "exposing_list",
  Field = "field",
  FieldAccessExpr = "field_access_expr",
  FieldAccessorFunctionExpr = "field_accessor_function_expr",
  FieldType = "field_type",
  File = "file",
  FunctionCallExpr = "function_call_expr",
  FunctionDeclarationLeft = "function_declaration_left",
  GlslCodeExpr = "glsl_code_expr",
  IfElseExpr = "if_else_expr",
  ImportClause = "import_clause",
  InfixDeclaration = "infix_declaration",
  LetInExpr = "let_in_expr",
  ListExpr = "list_expr",
  ListPattern = "list_pattern",
  LowerPattern = "lower_pattern",
  LowerTypeName = "lower_type_name",
  ModuleDeclaration = "module_declaration",
  NegateExpr = "negate_expr",
  NullaryConstructorArgumentPattern = "nullary_constructor_argument_pattern",
  NumberConstantExpr = "number_constant_expr",
  Operator = "operator",
  OperatorAsFunctionExpr = "operator_as_function_expr",
  OperatorIdentifier = "operator_identifier",
  ParenthesizedExpr = "parenthesized_expr",
  Pattern = "pattern",
  PortAnnotation = "port_annotation",
  RecordBaseIdentifier = "record_base_identifier",
  RecordExpr = "record_expr",
  RecordPattern = "record_pattern",
  RecordType = "record_type",
  StringConstantExpr = "string_constant_expr",
  TupleExpr = "tuple_expr",
  TuplePattern = "tuple_pattern",
  TupleType = "tuple_type",
  TypeAliasDeclaration = "type_alias_declaration",
  TypeAnnotation = "type_annotation",
  TypeDeclaration = "type_declaration",
  TypeExpression = "type_expression",
  TypeRef = "type_ref",
  TypeVariable = "type_variable",
  UnionPattern = "union_pattern",
  UnionVariant = "union_variant",
  UnitExpr = "unit_expr",
  UpperCaseQid = "upper_case_qid",
  ValueDeclaration = "value_declaration",
  ValueExpr = "value_expr",
  ValueQid = "value_qid",
  Alias = "alias",
  Arrow = "arrow",
  As = "as",
  Backslash = "backslash",
  Case = "case",
  CloseChar = "close_char",
  CloseQuote = "close_quote",
  Colon = "colon",
  DoubleDot = "double_dot",
  Effect = "effect",
  Eq = "eq",
  Exposing = "exposing",
  GlslContent = "glsl_content",
  Import = "import",
  Infix = "infix",
  InvalidStringEscape = "invalid_string_escape",
  LineComment = "line_comment",
  LowerCaseIdentifier = "lower_case_identifier",
  Module = "module",
  NumberLiteral = "number_literal",
  Of = "of",
  OpenChar = "open_char",
  OpenQuote = "open_quote",
  Port = "port",
  RegularStringPart = "regular_string_part",
  StringEscape = "string_escape",
  Type = "type",
  Underscore = "underscore",
  UpperCaseIdentifier = "upper_case_identifier",
  Where = "where",
}
// ...
export type SyntaxNode =
  | AnonymousFunctionExprNode
  | AnythingPatternNode
  | AsClauseNode
  | BinOpExprNode
  | BlockCommentNode
  | CaseOfBranchNode
  | CaseOfExprNode
  | CharConstantExprNode
  | ConsPatternNode
  | DotNode
  | ExposedOperatorNode
  | ExposedTypeNode
  | ExposedUnionConstructorsNode
  | ExposedValueNode
  | ExposingListNode
  | FieldNode
  | FieldAccessExprNode
  | FieldAccessorFunctionExprNode
  | FieldTypeNode
  | FileNode
  | FunctionCallExprNode
  | FunctionDeclarationLeftNode
  | GlslCodeExprNode
  | IfElseExprNode
  | ImportClauseNode
  | InfixDeclarationNode
  | LetInExprNode
  | ListExprNode
  | ListPatternNode
  | LowerPatternNode
  | LowerTypeNameNode
  | ModuleDeclarationNode
  | NegateExprNode
  | NullaryConstructorArgumentPatternNode
  | NumberConstantExprNode
  | OperatorNode
  | OperatorAsFunctionExprNode
  | OperatorIdentifierNode
  | ParenthesizedExprNode
  | PatternNode
  | PortAnnotationNode
  | RecordBaseIdentifierNode
  | RecordExprNode
  | RecordPatternNode
  | RecordTypeNode
  | StringConstantExprNode
  | TupleExprNode
  | TuplePatternNode
  | TupleTypeNode
  | TypeAliasDeclarationNode
  | TypeAnnotationNode
  | TypeDeclarationNode
  | TypeExpressionNode
  | TypeRefNode
  | TypeVariableNode
  | UnionPatternNode
  | UnionVariantNode
  | UnitExprNode
  | UpperCaseQidNode
  | ValueDeclarationNode
  | ValueExprNode
  | ValueQidNode
  | UnnamedNode<"&&">
  | UnnamedNode<"(">
  | UnnamedNode<")">
  | UnnamedNode<"*">
  | UnnamedNode<"+">
  | UnnamedNode<"++">
  | UnnamedNode<",">
  | UnnamedNode<"-">
  | UnnamedNode<"-}">
  | UnnamedNode<".">
  | UnnamedNode<"/">
  | UnnamedNode<"//">
  | UnnamedNode<"/=">
  | UnnamedNode<"::">
  | UnnamedNode<"<">
  | UnnamedNode<"</>">
  | UnnamedNode<"<<">
  | UnnamedNode<"<=">
  | UnnamedNode<"<?>">
  | UnnamedNode<"<|">
  | UnnamedNode<"==">
  | UnnamedNode<">">
  | UnnamedNode<">=">
  | UnnamedNode<">>">
  | UnnamedNode<"[">
  | UnnamedNode<"[glsl|">
  | UnnamedNode<"]">
  | UnnamedNode<"^">
  | AliasNode
  | ArrowNode
  | AsNode
  | BackslashNode
  | CaseNode
  | CloseCharNode
  | CloseQuoteNode
  | ColonNode
  | DoubleDotNode
  | EffectNode
  | UnnamedNode<"else">
  | EqNode
  | ExposingNode
  | GlslContentNode
  | UnnamedNode<"if">
  | ImportNode
  | UnnamedNode<"in">
  | InfixNode
  | InvalidStringEscapeNode
  | UnnamedNode<"let">
  | LineCommentNode
  | LowerCaseIdentifierNode
  | ModuleNode
  | NumberLiteralNode
  | OfNode
  | OpenCharNode
  | OpenQuoteNode
  | PortNode
  | RegularStringPartNode
  | StringEscapeNode
  | UnnamedNode<"then">
  | TypeNode
  | UnderscoreNode
  | UpperCaseIdentifierNode
  | WhereNode
  | UnnamedNode<"{">
  | UnnamedNode<"{-">
  | UnnamedNode<"|">
  | UnnamedNode<"|.">
  | UnnamedNode<"|=">
  | UnnamedNode<"|>">
  | UnnamedNode<"|]">
  | UnnamedNode<"||">
  | UnnamedNode<"}">
  | ErrorNode;

export interface AnonymousFunctionExprNode extends NamedNodeBase {
  type: SyntaxType.AnonymousFunctionExpr;
  exprNode:
    | AnonymousFunctionExprNode
    | BinOpExprNode
    | CaseOfExprNode
    | CharConstantExprNode
    | FieldAccessExprNode
    | FieldAccessorFunctionExprNode
    | FunctionCallExprNode
    | GlslCodeExprNode
    | IfElseExprNode
    | LetInExprNode
    | ListExprNode
    | NegateExprNode
    | NumberConstantExprNode
    | OperatorAsFunctionExprNode
    | ParenthesizedExprNode
    | RecordExprNode
    | StringConstantExprNode
    | TupleExprNode
    | UnitExprNode
    | ValueExprNode;
  paramNodes: PatternNode[];
}

export interface AnythingPatternNode extends NamedNodeBase {
  type: SyntaxType.AnythingPattern;
}

export interface AsClauseNode extends NamedNodeBase {
  type: SyntaxType.AsClause;
  nameNode: UpperCaseIdentifierNode;
}

export interface BinOpExprNode extends NamedNodeBase {
  type: SyntaxType.BinOpExpr;
  partNodes: (
    | AnonymousFunctionExprNode
    | CaseOfExprNode
    | CharConstantExprNode
    | FieldAccessExprNode
    | FieldAccessorFunctionExprNode
    | FunctionCallExprNode
    | GlslCodeExprNode
    | IfElseExprNode
    | LetInExprNode
    | ListExprNode
    | NegateExprNode
    | NumberConstantExprNode
    | OperatorNode
    | OperatorAsFunctionExprNode
    | ParenthesizedExprNode
    | RecordExprNode
    | StringConstantExprNode
    | TupleExprNode
    | UnitExprNode
    | ValueExprNode
  )[];
}

export interface BlockCommentNode extends NamedNodeBase {
  type: SyntaxType.BlockComment;
}

export interface CaseOfBranchNode extends NamedNodeBase {
  type: SyntaxType.CaseOfBranch;
  exprNode:
    | AnonymousFunctionExprNode
    | BinOpExprNode
    | CaseOfExprNode
    | CharConstantExprNode
    | FieldAccessExprNode
    | FieldAccessorFunctionExprNode
    | FunctionCallExprNode
    | GlslCodeExprNode
    | IfElseExprNode
    | LetInExprNode
    | ListExprNode
    | NegateExprNode
    | NumberConstantExprNode
    | OperatorAsFunctionExprNode
    | ParenthesizedExprNode
    | RecordExprNode
    | StringConstantExprNode
    | TupleExprNode
    | UnitExprNode
    | ValueExprNode;
  patternNode: PatternNode;
}

export interface CaseOfExprNode extends NamedNodeBase {
  type: SyntaxType.CaseOfExpr;
  branchNodes: CaseOfBranchNode[];
  exprNode:
    | AnonymousFunctionExprNode
    | BinOpExprNode
    | CaseOfExprNode
    | CharConstantExprNode
    | FieldAccessExprNode
    | FieldAccessorFunctionExprNode
    | FunctionCallExprNode
    | GlslCodeExprNode
    | IfElseExprNode
    | LetInExprNode
    | ListExprNode
    | NegateExprNode
    | NumberConstantExprNode
    | OperatorAsFunctionExprNode
    | ParenthesizedExprNode
    | RecordExprNode
    | StringConstantExprNode
    | TupleExprNode
    | UnitExprNode
    | ValueExprNode;
}

export interface CharConstantExprNode extends NamedNodeBase {
  type: SyntaxType.CharConstantExpr;
}

export interface ConsPatternNode extends NamedNodeBase {
  type: SyntaxType.ConsPattern;
  partNodes: (
    | UnnamedNode<"(">
    | UnnamedNode<")">
    | AnythingPatternNode
    | CharConstantExprNode
    | ConsPatternNode
    | ListPatternNode
    | LowerPatternNode
    | NumberConstantExprNode
    | PatternNode
    | RecordPatternNode
    | StringConstantExprNode
    | TuplePatternNode
    | UnionPatternNode
    | UnitExprNode
  )[];
}

export interface DotNode extends NamedNodeBase {
  type: SyntaxType.Dot;
}
// ...
export interface FileNode extends NamedNodeBase {
  type: SyntaxType.File;
  moduleDeclarationNode?: ModuleDeclarationNode;
}
// ...
export interface ModuleDeclarationNode extends NamedNodeBase {
  type: SyntaxType.ModuleDeclaration;
  exposingNode: ExposingListNode;
  nameNode: UpperCaseQidNode;
}

export interface NegateExprNode extends NamedNodeBase {
  type: SyntaxType.NegateExpr;
}

export interface NullaryConstructorArgumentPatternNode extends NamedNodeBase {
  type: SyntaxType.NullaryConstructorArgumentPattern;
}

export interface NumberConstantExprNode extends NamedNodeBase {
  type: SyntaxType.NumberConstantExpr;
}

export interface OperatorNode extends NamedNodeBase {
  type: SyntaxType.Operator;
}